scripts/requirements.txt
pillow==10.2.0
requests==2.31.0 